abstracted chart configurations away from components
