+++ src/components/Charts/mapChart.js
@@ -6,11 +6,11 @@ import Highcharts from 'highcharts';
 import HighchartsReact from 'highcharts-react-official';
 import highchartsMap from "highcharts/modules/map";
 import proj4 from "proj4";
-import mapDataCA from "@highcharts/map-collection/countries/us/us-ca-all.geo.json";
 import { Spinner, Icon, Dialog } from "@blueprintjs/core";
 
 import CustomMap from '../Map/index';
 import EventList from '../TrafficEvents/mapEventList';
+import { createMapChart } from './mapChartConfig';
 
 highchartsMap(Highcharts);
 
@@ -31,7 +31,9 @@ const MapChart = ({ events, data, filteredCounties }) => {
     // Create the chart on component mount
     useEffect(() => {
         if (events.length !== 0) {
-            createMapChart(events, data)
+            let options = createMapChart(events, data)
+            setMapOptions(options)
+            setLoading(false)
         }
     }, [events])
 
@@ -47,114 +49,21 @@ const MapChart = ({ events, data, filteredCounties }) => {
                     }
                 })
             })
-            createMapChart(cEvents, filteredCounties)
+            let options = createMapChart(cEvents, filteredCounties)
+            setMapOptions(options)
         }
     }, [filteredCounties])
 
-    // Highcharts map chart configuration
-    const createMapChart = (events, mapData) => {
-        // Filter out CONSTRUCTIONS for chart
-        const constructions = events.filter((ev) => {
-          return ev.keyword === 'CONSTRUCTION'
-        })
-
-        // Filter out INCIDENTS for chart
-        const incidents = events.filter((ev) => {
-          return ev.keyword === 'INCIDENT'
-        })
-
-        const mapOpt = {
-            chart: {
-                panning: true,
-                pinchType: 'x'
-            },
-            title: {
-              text: 'SF Bay Area'
-            },
-            credits: {
-                enabled: false
-            },
-            mapNavigation: {
-              enabled: true
-            },
-            tooltip: {
-                headerFormat: '',
-                pointFormat: 'County: {point.name}',
-                style: {
-                    fontWeight: 'bold'
-                }
-            },
-            series: [{
-              name: 'SF Bay map',
-              mapData: mapDataCA,
-              borderColor: '#A0A0A0',
-              color: '#a0aec0',
-              nullColor: 'rgba(200, 200, 200, 0.3)',
-              showInLegend: false,
-              data: mapData,
-              allAreas: false,
-              joinBy: 'hc-key',
-              allowPointSelect: true,
-                cursor: 'pointer',
-                events: {
-                    click: function (e) {
-                        e.point.zoomTo();
-                    }
-                },
-              dataLabels: {
-                    enabled: true,
-                    format: '{point.hc-a2}'
-                }
-            }, {
-              // Specify points using lat/lon
-              type: 'mappoint',
-              name: 'Constructions',
-              color: '#ed8936',
-              data: constructions,
-              dataLabels: {
-                enabled: true,
-                format: '{point.keyword}',
-                style: {
-                    color: '#ed8936'
-                }
-              },
-              tooltip: {
-                headerFormat: '',
-                pointFormat: `<div><span>{point.keyword}</span><br><span>Status: {point.status}</span> <br><span>State: {point.roadsState}</span> <br><span>From: {point.from}</span> <br><span>To: {point.to}</span> <br><span>Direction: {point.direction}</span> <br><span>Updated: {point.updated}</span></div>`
-              },
-              cursor: 'pointer',
-            }, {
-              // Specify points using lat/lon
-              type: 'mappoint',
-              name: 'Incidents',
-              color: '#e53e3e',
-              data: incidents,
-              dataLabels: {
-                enabled: true,
-                format: '{point.keyword}',
-                style: {
-                    color: '#e53e3e'
-                }
-              },
-              tooltip: {
-                headerFormat: '',
-                pointFormat: `<div><span>{point.keyword}</span><br><span>Status: {point.status}</span> <br><span>State: {point.roadsState}</span> <br><span>From: {point.from}</span> <br><span>To: {point.to}</span> <br><span>Direction: {point.direction}</span> <br><span>Updated: {point.updated}</span></div>`
-              },
-              cursor: 'pointer',
-            }]
-        }
-
-        setMapOptions(mapOpt)
-        setLoading(false)
-    }
-
+    // Clear events from map and only display marker for the selected event
     const showEventOnMap = (id) => {
         const filteredList = events.filter((e) => {
             return e.id === id
         })
-        createMapChart(filteredList, data)
+        let options = createMapChart(filteredList, data)
+        setMapOptions(options)
     }
 
+    // Select a specific event -> set it's state to 'selected' so it changes visual appearance
     const selectEvent = (index) => {
         for (let i = 0; i < selected.length; i++){
             selected[i] = false
@@ -190,6 +99,7 @@ const MapChart = ({ events, data, filteredCounties }) => {
                   createMapChart={createMapChart}
                   isVisible={isVisible}
                   data={data}
+                  setMapOptions={setMapOptions}
                 />
                 <Dialog
                     isOpen={isOpen}
