+++ src/components/TrafficEvents/mapEventList.js
@@ -6,13 +6,13 @@ import { Icon } from "@blueprintjs/core";
 
 import EventListItem from '../TrafficEvents/mapEventListItem';
 
-const EventList = ({ data, events, showEventOnMap, selectEvent, selected, setSelected, setIsVisible, setIsOpen, setCoordinates, setMapData, isVisible, createMapChart }) => {
+const EventList = ({ data, events, showEventOnMap, selectEvent, selected, setSelected, setIsVisible, setIsOpen, setCoordinates, setMapData, isVisible, createMapChart, setMapOptions }) => {
     return (
         <div className="flex flex-col md:w-1/3 md:l-0 md:shadow-lg md:mr-auto md:-mt-12">
             <div className="w-full flex flex-row justify-between items-center bg-blue-700 text-white font-semibold p-4 shadow-lg">
                 <p className="text-lg">Latest events</p>
                 <span className={isVisible ? 'hidden' : `text-sm text-gray-400`}>total: {events.length}</span>
-                <button className={!isVisible ? 'hidden' : 'block'} onClick={() => {createMapChart(events, data); selectEvent(); setIsVisible(false)}}><Icon icon="reset" size="15" className="text-white"/></button>
+                <button className={!isVisible ? 'hidden' : 'block'} onClick={() => {setMapOptions(createMapChart(events, data)); selectEvent(); setIsVisible(false)}}><Icon icon="reset" size="15" className="text-white"/></button>
             </div>
             <div className="w-full text-xs">
                 <ul className="m-auto overflow-auto shadow-inner event-list-for-map">
